protobuf decoding byte analysis
test to try and find what bytes change in the base64 encoded data string sent from the eufy api to see if a match can be found with the android app accessory usage levels then a reset was performed in the android app and the accessory base64 data string was analyzed looking for changes to each item one at a time. from these changes Home Assistant sensors were created to be used to monitor the usage and automations can be used to send warnings to the user thru H.A or a mobile phone etc etc.